wip: colocation refactor
