+++ src/app/page.tsx
@@ -2,7 +2,7 @@ import { Graph } from '@/components/graph'
 import { Library } from '@/components/library'
 import { Output } from '@/components/output'
 import { Properties } from '@/components/properties'
-import styles from '@/styles/app.page.module.css'
+import styles from './app.page.module.css'
 
 export default function Home() {
   return (
