+++ src/components/node/node.module.css
@@ -35,26 +35,12 @@
   cursor: grabbing;
 }
 
-.slider {
-  cursor: pointer;
-  width: 100%;
-}
-
-.slider:active {
-  cursor: move;
-}
-
 .title {
   margin: 0;
   width: 100%;
   font-weight: 800;
   font-size: medium;
-  text-align-last: center;
-}
-
-.canvas {
-  width: 100%;
-  height: 100%;
+  text-align: center;
 }
 
 .inputs,
