+++ src/components/port/port.tsx
@@ -1,10 +1,27 @@
 'use client'
 
-import { useGraph } from '@/contexts/graph.povider'
-import { GraphActionTypes } from '@/enums/graph'
-import styles from '@/styles/port.module.css'
-import type { PortProps } from '@/types/port'
 import { useEffect, useRef, type MouseEvent } from 'react'
+import { GraphActionTypes, useGraph } from '../graph'
+import styles from './port.module.css'
+
+export enum PortKind {
+  Input = 'input',
+  Output = 'output',
+}
+
+export enum PortStatus {
+  Idle = 'idle',
+  Active = 'active',
+  Linked = 'linked',
+}
+
+export interface PortProps {
+  nodeId?: string | undefined
+  id: string | undefined
+  value: number | boolean | string | undefined
+  kind: PortKind
+  status: PortStatus
+}
 
 export function Port({ id, value, nodeId }: PortProps) {
   const ref = useRef<HTMLButtonElement>(null)
