+++ src/components/properties/properties.module.css
