+++ src/components/stream/stream.tsx
@@ -1,9 +1,24 @@
 'use client'
 
-import { useGraph } from '@/contexts/graph.povider'
-import { GraphActionTypes } from '@/enums/graph'
-import type { StreamProps } from '@/types/stream'
 import { useEffect } from 'react'
+import { GraphActionTypes, useGraph } from '../graph'
+
+export enum StreamStatus {
+  Active = 'active',
+  Linked = 'linked',
+}
+
+export interface StreamProps {
+  id: string | undefined
+  value: number | boolean | string | undefined
+  m: string
+  l?: string
+  status: StreamStatus
+  sourceId: string | undefined
+  targetId?: string | undefined
+  source: HTMLButtonElement
+  target: HTMLButtonElement | null
+}
 
 export function Stream({ m, l, value, targetId }: StreamProps) {
   const { dispatch } = useGraph()
