--- src/contexts/graph.reducer.ts
@@ -1,887 +1,0 @@
-import { GraphActionTypes } from '@/enums/graph'
-import { NodeKind, NodeStatus, NodeVariant } from '@/enums/node'
-import { PortKind, PortStatus } from '@/enums/port'
-import { StreamStatus } from '@/enums/stream'
-import { GraphAction, GraphState } from '@/types/graph'
-
-export function graphsReducer(
-  state: GraphState,
-  action: GraphAction
-): GraphState {
-  const gap = 18
-  const nodeHeight = 128
-
-  switch (action.type) {
-    case GraphActionTypes.GRAPH_MOUSE_MOVE: {
-      const { nodes, streams } = state
-      const { clientX, clientY } = action.payload.event
-
-      if (
-        nodes.every((node) => node.status !== NodeStatus.Active) &&
-        streams.every((stream) => stream.status !== StreamStatus.Active)
-      ) {
-        return state
-      }
-
-      return {
-        ...state,
-        nodes: nodes.map((node) => {
-          const { x, y } = node.offset
-          if (node.status !== NodeStatus.Active) return node
-
-          return {
-            ...node,
-            position: {
-              x: clientX - x,
-              y: clientY - y,
-            },
-          }
-        }),
-        streams: streams.map((stream) => {
-          const { source, target } = stream
-          const sourceBounds = source.getBoundingClientRect()
-          const targetBounds = target?.getBoundingClientRect()
-          const sourceX = sourceBounds.x + sourceBounds.width * 0.5 - gap
-          const sourceY = sourceBounds.y + sourceBounds.height * 0.5 - gap
-
-          if (stream.status === StreamStatus.Linked && targetBounds) {
-            const targetX = targetBounds.x + targetBounds.width * 0.5 - gap
-            const targetY = targetBounds.y + targetBounds.height * 0.5 - gap
-
-            return {
-              ...stream,
-              m: `${sourceX} ${sourceY}`,
-              l: `${targetX} ${targetY}`,
-            }
-          }
-
-          if (stream.status !== StreamStatus.Active) {
-            return stream
-          }
-
-          return {
-            ...stream,
-            l: `${clientX - gap} ${clientY - gap}`,
-          }
-        }),
-      }
-    }
-    case GraphActionTypes.GRAPH_MOUSE_UP: {
-      const { nodes, streams } = state
-      if (
-        nodes.every((node) => node.status !== NodeStatus.Active) &&
-        streams.every((stream) => stream.status !== StreamStatus.Active)
-      ) {
-        return state
-      }
-
-      return {
-        ...state,
-        nodes: nodes.map((node) => {
-          return {
-            ...node,
-            status: NodeStatus.Idle,
-            ports: node.ports.map((port) => {
-              if (
-                port.id !==
-                streams.find((stream) => stream.status === StreamStatus.Active)
-                  ?.sourceId
-              )
-                return port
-              return {
-                ...port,
-                status: PortStatus.Idle,
-              }
-            }),
-          }
-        }),
-        streams: streams
-          .map((stream) => {
-            const { source, target } = stream
-            const sourceBounds = source.getBoundingClientRect()
-            const targetBounds = target?.getBoundingClientRect()
-            const sourceX = sourceBounds.x + sourceBounds.width * 0.5 - gap
-            const sourceY = sourceBounds.y + sourceBounds.height * 0.5 - gap
-            const targetX =
-              targetBounds && targetBounds?.x + targetBounds.width * 0.5 - gap
-            const targetY =
-              targetBounds && targetBounds?.y + targetBounds.height * 0.5 - gap
-            return {
-              ...stream,
-              m: `${sourceX} ${sourceY}`,
-              l: `${targetX} ${targetY}`,
-            }
-          })
-          .filter((stream) => stream.status === StreamStatus.Linked),
-      }
-    }
-    case GraphActionTypes.GRAPH_MOUSE_LEAVE: {
-      const { streams } = state
-      return {
-        ...state,
-
-        streams: streams
-          .map((stream) => {
-            const { source, target } = stream
-            const sourceBounds = source.getBoundingClientRect()
-            const targetBounds = target?.getBoundingClientRect()
-            const sourceX = sourceBounds.x + sourceBounds.width * 0.5 - gap
-            const sourceY = sourceBounds.y + sourceBounds.height * 0.5 - gap
-            const targetX =
-              targetBounds && targetBounds?.x + targetBounds.width * 0.5 - gap
-            const targetY =
-              targetBounds && targetBounds?.y + targetBounds.height * 0.5 - gap
-            return {
-              ...stream,
-              m: `${sourceX} ${sourceY}`,
-              l: `${targetX} ${targetY}`,
-            }
-          })
-          .filter((stream) => stream.status === StreamStatus.Linked),
-      }
-    }
-    case GraphActionTypes.GRAPH_DROP: {
-      const { nodes } = state
-      const { clientX, clientY } = action.payload.event
-      const center = nodeHeight * 0.5
-
-      switch (action.payload.event.dataTransfer.getData('node')) {
-        case NodeVariant.Integer: {
-          return {
-            ...state,
-            nodes: [
-              ...nodes,
-              {
-                id: String(nodes.length + 1),
-                kind: NodeKind.Input,
-                variant: NodeVariant.Integer,
-                position: {
-                  x: clientX - gap - center,
-                  y: clientY - gap - center,
-                },
-                value: 0,
-                offset: { x: 0, y: 0 },
-                status: NodeStatus.Idle,
-                ports: [
-                  {
-                    id: `${nodes.length + 1}-1`,
-                    kind: PortKind.Output,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                ],
-              },
-            ],
-          }
-        }
-        case NodeVariant.Boolean: {
-          return {
-            ...state,
-            nodes: [
-              ...nodes,
-              {
-                id: String(nodes.length + 1),
-                kind: NodeKind.Input,
-                variant: NodeVariant.Boolean,
-                position: {
-                  x: clientX - gap - center,
-                  y: clientY - gap - center,
-                },
-                value: false,
-                offset: { x: 0, y: 0 },
-                status: NodeStatus.Idle,
-                ports: [
-                  {
-                    id: `${nodes.length + 1}-1`,
-                    kind: PortKind.Output,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                ],
-              },
-            ],
-          }
-        }
-        case NodeVariant.Float: {
-          return {
-            ...state,
-            nodes: [
-              ...nodes,
-              {
-                id: String(nodes.length + 1),
-                kind: NodeKind.Input,
-                variant: NodeVariant.Float,
-                position: {
-                  x: clientX - gap - center,
-                  y: clientY - gap - center,
-                },
-                value: 0.0,
-                offset: { x: 0, y: 0 },
-                status: NodeStatus.Idle,
-                ports: [
-                  {
-                    id: `${nodes.length + 1}-1`,
-                    kind: PortKind.Output,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                ],
-              },
-            ],
-          }
-        }
-        case NodeVariant.String: {
-          return {
-            ...state,
-            nodes: [
-              ...nodes,
-              {
-                id: String(nodes.length + 1),
-                kind: NodeKind.Input,
-                variant: NodeVariant.String,
-                position: {
-                  x: clientX - gap - center,
-                  y: clientY - gap - center,
-                },
-                value: '',
-                offset: { x: 0, y: 0 },
-                status: NodeStatus.Idle,
-                ports: [
-                  {
-                    id: `${nodes.length + 1}-1`,
-                    kind: PortKind.Output,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                ],
-              },
-            ],
-          }
-        }
-        case NodeVariant.Addition: {
-          return {
-            ...state,
-            nodes: [
-              ...nodes,
-              {
-                id: String(nodes.length + 1),
-                kind: NodeKind.Opertator,
-                variant: NodeVariant.Addition,
-                position: {
-                  x: clientX - gap - center,
-                  y: clientY - gap - center,
-                },
-                value: undefined,
-                offset: { x: 0, y: 0 },
-                status: NodeStatus.Idle,
-                ports: [
-                  {
-                    id: `${nodes.length + 1}-1`,
-                    kind: PortKind.Input,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                  {
-                    id: `${nodes.length + 1}-2`,
-                    kind: PortKind.Input,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                  {
-                    id: `${nodes.length + 1}-3`,
-                    kind: PortKind.Output,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                ],
-              },
-            ],
-          }
-        }
-        case NodeVariant.Subtraction: {
-          return {
-            ...state,
-            nodes: [
-              ...nodes,
-              {
-                id: String(nodes.length + 1),
-                kind: NodeKind.Opertator,
-                variant: NodeVariant.Subtraction,
-                position: {
-                  x: clientX - gap - center,
-                  y: clientY - gap - center,
-                },
-                value: undefined,
-                offset: { x: 0, y: 0 },
-                status: NodeStatus.Idle,
-                ports: [
-                  {
-                    id: `${nodes.length + 1}-1`,
-                    kind: PortKind.Input,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                  {
-                    id: `${nodes.length + 1}-2`,
-                    kind: PortKind.Input,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                  {
-                    id: `${nodes.length + 1}-3`,
-                    kind: PortKind.Output,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                ],
-              },
-            ],
-          }
-        }
-        case NodeVariant.Division: {
-          return {
-            ...state,
-            nodes: [
-              ...nodes,
-              {
-                id: String(nodes.length + 1),
-                kind: NodeKind.Opertator,
-                variant: NodeVariant.Division,
-                position: {
-                  x: clientX - gap - center,
-                  y: clientY - gap - center,
-                },
-                value: undefined,
-                offset: { x: 0, y: 0 },
-                status: NodeStatus.Idle,
-                ports: [
-                  {
-                    id: `${nodes.length + 1}-1`,
-                    kind: PortKind.Input,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                  {
-                    id: `${nodes.length + 1}-2`,
-                    kind: PortKind.Input,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                  {
-                    id: `${nodes.length + 1}-3`,
-                    kind: PortKind.Output,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                ],
-              },
-            ],
-          }
-        }
-        case NodeVariant.Multiplication: {
-          return {
-            ...state,
-            nodes: [
-              ...nodes,
-              {
-                id: String(nodes.length + 1),
-                kind: NodeKind.Opertator,
-                variant: NodeVariant.Multiplication,
-                position: {
-                  x: clientX - gap - center,
-                  y: clientY - gap - center,
-                },
-                value: undefined,
-                offset: { x: 0, y: 0 },
-                status: NodeStatus.Idle,
-                ports: [
-                  {
-                    id: `${nodes.length + 1}-1`,
-                    kind: PortKind.Input,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                  {
-                    id: `${nodes.length + 1}-2`,
-                    kind: PortKind.Input,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                  {
-                    id: `${nodes.length + 1}-3`,
-                    kind: PortKind.Output,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                ],
-              },
-            ],
-          }
-        }
-        case NodeVariant.Modulo: {
-          return {
-            ...state,
-            nodes: [
-              ...nodes,
-              {
-                id: String(nodes.length + 1),
-                kind: NodeKind.Opertator,
-                variant: NodeVariant.Modulo,
-                position: {
-                  x: clientX - gap - center,
-                  y: clientY - gap - center,
-                },
-                value: undefined,
-                offset: { x: 0, y: 0 },
-                status: NodeStatus.Idle,
-                ports: [
-                  {
-                    id: `${nodes.length + 1}-1`,
-                    kind: PortKind.Input,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                  {
-                    id: `${nodes.length + 1}-2`,
-                    kind: PortKind.Input,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                  {
-                    id: `${nodes.length + 1}-3`,
-                    kind: PortKind.Output,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                ],
-              },
-            ],
-          }
-        }
-        case NodeVariant.Exponentiation: {
-          return {
-            ...state,
-            nodes: [
-              ...nodes,
-              {
-                id: String(nodes.length + 1),
-                kind: NodeKind.Opertator,
-                variant: NodeVariant.Exponentiation,
-                position: {
-                  x: clientX - gap - center,
-                  y: clientY - gap - center,
-                },
-                value: undefined,
-                offset: { x: 0, y: 0 },
-                status: NodeStatus.Idle,
-                ports: [
-                  {
-                    id: `${nodes.length + 1}-1`,
-                    kind: PortKind.Input,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                  {
-                    id: `${nodes.length + 1}-2`,
-                    kind: PortKind.Input,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                  {
-                    id: `${nodes.length + 1}-3`,
-                    kind: PortKind.Output,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                ],
-              },
-            ],
-          }
-        }
-        case NodeVariant.Result: {
-          return {
-            ...state,
-            nodes: [
-              ...nodes,
-              {
-                id: String(nodes.length + 1),
-                kind: NodeKind.Output,
-                variant: NodeVariant.Result,
-                position: {
-                  x: clientX - gap - center,
-                  y: clientY - gap - center,
-                },
-                value: undefined,
-                offset: { x: 0, y: 0 },
-                status: NodeStatus.Idle,
-                ports: [
-                  {
-                    id: `${nodes.length + 1}-1`,
-                    kind: PortKind.Input,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                ],
-              },
-            ],
-          }
-        }
-        case NodeVariant.Color: {
-          return {
-            ...state,
-            nodes: [
-              ...nodes,
-              {
-                id: String(nodes.length + 1),
-                kind: NodeKind.Input,
-                variant: NodeVariant.Color,
-                position: {
-                  x: clientX - gap - center,
-                  y: clientY - gap - center,
-                },
-                value: undefined,
-                offset: { x: 0, y: 0 },
-                status: NodeStatus.Idle,
-                ports: [
-                  {
-                    id: `${nodes.length + 1}-1`,
-                    kind: PortKind.Output,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                ],
-              },
-            ],
-          }
-        }
-        case NodeVariant.WebGpu: {
-          return {
-            ...state,
-            nodes: [
-              ...nodes,
-              {
-                id: String(nodes.length + 1),
-                kind: NodeKind.Input,
-                variant: NodeVariant.WebGpu,
-                position: {
-                  x: clientX - gap - center,
-                  y: clientY - gap - center,
-                },
-                value: undefined,
-                offset: { x: 0, y: 0 },
-                status: NodeStatus.Idle,
-                ports: [
-                  {
-                    id: `${nodes.length + 1}-1`,
-                    kind: PortKind.Output,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                ],
-              },
-            ],
-          }
-        }
-        case NodeVariant.Blend: {
-          return {
-            ...state,
-            nodes: [
-              ...nodes,
-              {
-                id: String(nodes.length + 1),
-                kind: NodeKind.Opertator,
-                variant: NodeVariant.Blend,
-                position: {
-                  x: clientX - gap - center,
-                  y: clientY - gap - center,
-                },
-                value: undefined,
-                offset: { x: 0, y: 0 },
-                status: NodeStatus.Idle,
-                ports: [
-                  {
-                    id: `${nodes.length + 1}-1`,
-                    kind: PortKind.Input,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                  {
-                    id: `${nodes.length + 1}-2`,
-                    kind: PortKind.Input,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                  {
-                    id: `${nodes.length + 1}-3`,
-                    kind: PortKind.Output,
-                    value: undefined,
-                    status: PortStatus.Idle,
-                  },
-                ],
-              },
-            ],
-          }
-        }
-      }
-      return state
-    }
-    case GraphActionTypes.NODE_MOUSE_DOWN: {
-      const { nodes } = state
-      const { clientX, clientY } = action.payload.event
-      const { id } = action.payload
-
-      return {
-        ...state,
-        nodes: nodes.map((node) => {
-          const { x, y } = node.position
-          if (node.id !== id) return node
-
-          return {
-            ...node,
-            offset: {
-              x: clientX - x,
-              y: clientY - y,
-            },
-            status: NodeStatus.Active,
-            position: {
-              x: node.position.x,
-              y: node.position.y,
-            },
-          }
-        }),
-      }
-    }
-    case GraphActionTypes.NODE_CLICK: {
-      const { nodes } = state
-      const { id } = action.payload
-
-      return {
-        ...state,
-        nodes: nodes.map((node) => {
-          return {
-            ...node,
-            selected: node.id === id ? true : false,
-            value: node.id === id ? node.value : node.value,
-          }
-        }),
-      }
-    }
-    case GraphActionTypes.PORT_MOUSE_DOWN: {
-      const { streams, nodes } = state
-      const { value, ref, id } = action.payload
-
-      if (!ref.current) throw new Error('Invalid port reference')
-      const { x, y, width, height } = ref.current.getBoundingClientRect()
-
-      return {
-        ...state,
-        nodes: nodes.map((node) => {
-          return {
-            ...node,
-            ports: node.ports.map((port) => {
-              if (port.id !== id) return port
-
-              return {
-                ...port,
-                status: PortStatus.Active,
-              }
-            }),
-          }
-        }),
-        streams: [
-          ...streams,
-          {
-            id: String(streams.length + 1),
-            value: value,
-            m: `${x + width * 0.5 - gap} ${y + height * 0.5 - gap}`,
-            l: `${x + width * 0.5 - gap} ${y + height * 0.5 - gap}`,
-            status: StreamStatus.Active,
-            sourceId: id,
-            source: ref.current,
-            target: null,
-          },
-        ],
-      }
-    }
-    case GraphActionTypes.PORT_MOUSE_UP: {
-      const { nodes, streams } = state
-      const { ref, id } = action.payload
-
-      if (!ref.current) throw new Error('Invalid port reference')
-      const { x, y, width, height } = ref.current.getBoundingClientRect()
-
-      return {
-        ...state,
-        nodes: nodes.map((node) => {
-          return {
-            ...node,
-            ports: node.ports.map((port) => {
-              if (
-                streams.find((stream) => stream.status === StreamStatus.Active)
-                  ?.sourceId &&
-                port.id !== id
-              )
-                return port
-
-              return {
-                ...port,
-                status: PortStatus.Linked,
-                value: streams.find(
-                  (stream) => stream.status === StreamStatus.Active
-                )?.value,
-              }
-            }),
-          }
-        }),
-        streams: streams.map((stream) => {
-          if (stream.status !== StreamStatus.Active) return stream
-
-          return {
-            ...stream,
-            status: StreamStatus.Linked,
-            l: `${x + width * 0.5 - gap} ${y + height * 0.5 - gap}`,
-            target: ref.current,
-            targetId: id,
-          }
-        }),
-      }
-    }
-    case GraphActionTypes.NODE_SLIDER_CHANGE: {
-      const { nodes } = state
-      const { id, event } = action.payload
-      const { value } = event.target as HTMLInputElement
-
-      return {
-        ...state,
-        nodes: nodes.map((node) => {
-          if (node.id !== id) return node
-
-          return {
-            ...node,
-            value:
-              node.variant === NodeVariant.Boolean
-                ? event.target.checked
-                  ? true
-                  : false
-                : node.variant === NodeVariant.String
-                ? event.target.value
-                : Number(value),
-          }
-        }),
-      }
-    }
-    case GraphActionTypes.NODE_VALUE_CHANGE: {
-      const { nodes } = state
-      const { value, id } = action.payload
-
-      return {
-        ...state,
-        nodes: nodes.map((node) => {
-          if (node.id !== id) return node
-
-          return {
-            ...node,
-            ports: node.ports.map((port) => {
-              if (port.kind !== PortKind.Output) return port
-
-              return {
-                ...port,
-                value: value,
-              }
-            }),
-          }
-        }),
-      }
-    }
-    case GraphActionTypes.PORT_VALUE_CHANGE: {
-      const { nodes, streams } = state
-      const { value, id, nodeId } = action.payload
-
-      return {
-        ...state,
-        nodes: nodes.map((node) => {
-          if (node.id !== nodeId) return node
-          return {
-            ...node,
-            value:
-              node.ports[0].value && node.variant === NodeVariant.Result
-                ? Number(node.ports[0].value)
-                : node.ports[0].value &&
-                  node.ports[1].value &&
-                  node.variant === NodeVariant.Addition
-                ? Number(node.ports[0].value) + Number(node.ports[1].value)
-                : node.ports[0].value &&
-                  node.ports[1].value &&
-                  node.variant === NodeVariant.Subtraction
-                ? Number(node.ports[0].value) - Number(node.ports[1].value)
-                : node.ports[0].value &&
-                  node.ports[1].value &&
-                  node.variant === NodeVariant.Multiplication
-                ? Number(node.ports[0].value) * Number(node.ports[1].value)
-                : node.ports[0].value &&
-                  node.ports[1].value &&
-                  node.variant === NodeVariant.Division
-                ? Number(node.ports[0].value) / Number(node.ports[1].value)
-                : node.ports[0].value &&
-                  node.ports[1].value &&
-                  node.variant === NodeVariant.Modulo
-                ? Number(node.ports[0].value) % Number(node.ports[1].value)
-                : node.ports[0].value &&
-                  node.ports[1].value &&
-                  node.variant === NodeVariant.Exponentiation
-                ? Number(node.ports[0].value) ** Number(node.ports[1].value)
-                : value,
-          }
-        }),
-        streams: streams.map((stream) => {
-          if (stream.sourceId !== id) return stream
-
-          return {
-            ...stream,
-            value: value,
-          }
-        }),
-      }
-    }
-    case GraphActionTypes.STREAM_VALUE_CHANGE: {
-      const { nodes } = state
-      const { value, targetId } = action.payload
-
-      return {
-        ...state,
-        nodes: nodes.map((node) => {
-          return {
-            ...node,
-            ports: node.ports.map((port) => {
-              if (port.id !== targetId) return port
-
-              return {
-                ...port,
-                value: value,
-              }
-            }),
-          }
-        }),
-      }
-    }
-    case GraphActionTypes.NODE_COLOR_CHANGE: {
-      const { nodes } = state
-      const { id, event } = action.payload
-      const { value } = event.target as HTMLInputElement
-
-      return {
-        ...state,
-        nodes: nodes.map((node) => {
-          if (node.id !== id) return node
-
-          return {
-            ...node,
-            value: value,
-          }
-        }),
-      }
-    }
-    default:
-      throw new Error(`Unhandled action type: ${GraphActionTypes}`)
-  }
-}
