--- src/data/library.ts
@@ -1,18 +1,0 @@
-import { NodeVariant } from '@/enums/node'
-
-export const nodeData = [
-  NodeVariant.Addition,
-  NodeVariant.Blend,
-  NodeVariant.Boolean,
-  NodeVariant.Color,
-  NodeVariant.Division,
-  NodeVariant.Exponentiation,
-  NodeVariant.Float,
-  NodeVariant.Integer,
-  NodeVariant.Modulo,
-  NodeVariant.Multiplication,
-  NodeVariant.Result,
-  NodeVariant.String,
-  NodeVariant.Subtraction,
-  NodeVariant.WebGpu,
-]
