--- src/enums/graph.ts
@@ -1,15 +1,0 @@
-export enum GraphActionTypes {
-  GRAPH_MOUSE_MOVE = 'graph_mouse_move',
-  GRAPH_MOUSE_UP = 'graph_mouse_up',
-  GRAPH_MOUSE_LEAVE = 'graph_mouse_leave',
-  GRAPH_DROP = 'graph_drop',
-  NODE_CLICK = 'node_click',
-  NODE_MOUSE_DOWN = 'node_mouse_down',
-  PORT_MOUSE_DOWN = 'port_mouse_down',
-  PORT_MOUSE_UP = 'port_mouse_up',
-  NODE_SLIDER_CHANGE = 'node_slider_change',
-  NODE_COLOR_CHANGE = 'node_color_change',
-  NODE_VALUE_CHANGE = 'node_value_change',
-  PORT_VALUE_CHANGE = 'port_value_change',
-  STREAM_VALUE_CHANGE = 'stream_value_change',
-}
