--- src/enums/port.ts
@@ -1,10 +1,0 @@
-export enum PortKind {
-  Input = 'input',
-  Output = 'output',
-}
-
-export enum PortStatus {
-  Idle = 'idle',
-  Active = 'active',
-  Linked = 'linked',
-}
