--- src/enums/stream.ts
@@ -1,4 +1,0 @@
-export enum StreamStatus {
-  Active = 'active',
-  Linked = 'linked',
-}
