--- src/types/root.layout.types.ts
@@ -1,5 +1,0 @@
-import type { ReactNode } from 'react'
-
-export interface RootLayoutProps {
-  children: ReactNode
-}
