+++ tsconfig.json
@@ -17,8 +17,7 @@
     "incremental": true,
     "baseUrl": ".",
     "paths": {
-      "@/*": ["src/*"],
-      "@/styles/*": ["src/styles/*"]
+      "@/*": ["src/*"]
     },
     "plugins": [
       {
